requirements.txt
Django==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.1.0
openpyxl==3.1.2
reportlab==4.0.7
python-decouple==3.8
pandas==2.1.4
xlrd==2.0.1
pymongo==4.8.0
bcrypt==4.1.2
gunicorn==21.2.0